Refactored folder setup
