+++ backend/prisma/prisma.yml
@@ -1,4 +1,4 @@
-endpoint: http://35.246.149.86
+endpoint: http://35.246.149.86:4466
 datamodel: datamodel.prisma
 generate:
   - generator: typescript-client
