+++ backend/src/utils/customTypes.ts
@@ -1,0 +1,6 @@
+import { ContextParameters } from 'graphql-yoga/dist/types';
+
+export interface Context {
+    request: ContextParameters;
+    user?: { name: string; email: string; role: string };
+}
